refactor(actions): extract portalModal action from HeroImageDisplay
HeroImageDisplay had an inline portalWithEvents action (portal to
document.body + click/keydown(Escape) dismissal + autofocus) that is
genuinely reusable for any modal that *is* its own backdrop. Move it
to $lib/actions/portalModal.ts with a small typed options API
({ onDismiss, target?, autoFocus? }). HeroImageDisplay now uses
<div use:portalModal={{ onDismiss: closeZoom }} ...>; behavior is
unchanged.
